test-minions-observe-llm/promptSim/app/requirements.txt
locust>=2.15.0
python-dotenv